Updated the tests based on numerical precision.
Sometimes transforming between bins, values, and back doesn't yield the
starting value. The tests now account for this.

This still seems a little weird, but I think it's due to imprecise
floating point representations for certain values. So, it's not really a
flaw in the code, but a flaw in the representation. And I don't really
think it matters.