Fixed a race condition that caused a deadlock on ctrl-C
When ctrl-C was pressed at just the right moment, the main thread would
be stuck in processAccumulatedFrames waiting for the accumulator to
have more than 20 frames in it. However, if the reader was also
blocked at that moment, neither could notice that the merge was
aborted, so they would both wait forever. This was fixed by blocking
the main thread before processAccumulatedFrames and including
abortWasCalled in the conditions for unblocking.